pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "fastapi logging with elk stack"
authors = ["Quốc Việt <thaibinhquocviet2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pyyaml = "^6.0.2"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
